+++ psgscoring/__init__.py
@@ -90,7 +90,7 @@ from .utils import (
     safe_r,
 )
 
-__version__ = "0.10.0"
+__version__ = "0.11.0"
 __all__ = [
     # Master
     "run_pneumo_analysis",